+++ apps/backoffice-v2/src/common/hooks/useFilterId/useFilterId.tsx
@@ -2,7 +2,6 @@ import { useSearchParams } from 'react-router-dom';
 
 export const useFilterId = (): string | undefined => {
   const [params] = useSearchParams();
-  const filterId = params.get('filterId');
 
-  return filterId;
+  return params.get('filterId');
 };