+++ apps/backoffice-v2/src/pages/Entity/components/CallToAction/CallToAction.tsx
@@ -248,9 +248,8 @@ export const CallToAction: FunctionComponent<ICallToActionProps> = ({ value, dat
                   reason: comment ? `${reason} - ${comment}` : reason,
                 })}
               >
-                {workflowLevelResolution ? (
-                  'Approve'
-                ) : (
+                {workflowLevelResolution && 'Confirm'}
+                {!workflowLevelResolution && (
                   <>
                     <Send size={18} />
                     Send email